Atomically initialize BoltDB
This fixes several possible crash issues found by ALICE testing.

BoltDB automatically creates itself on first open but doesn't do so in a
crash-safe way:
 - It creates a new file, appends zeroes up to the initial size and then
   writes a meta page. Crash between any of these leave the file in an
   invalid state that causes an error on startup.
 - It never calls fsync on the parent dir so even if succesful and later
   commits are made, the file may be lost on a crash because the Dir Entry
   wasn't persistent yet.

We fix this by doing init on a tmp file name and only renaming and then
fsynching the parent dir once the file is in a known-good state.